# openapi.json
{
  "components": {
    "schemas": {
...
      "AffectedPlayerRating": {
        "properties": {
          "discord_id": {
            "title": "Discord Id",
            "type": "string"
          },
          "rating_mu": {
            "title": "Rating Mu",
            "type": "number"
          }
        },
        "required": [
          "discord_id",
          "rating_mu"
        ],
        "title": "AffectedPlayerRating",
        "type": "object"
      },
...
      "Body_upload_match_api_v2_matches_post": {
        "properties": {
          "discord_message_id": {
            "title": "Discord Message Id",
            "type": "string"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "is_cloud": {
            "title": "Is Cloud",
            "type": "string"
          }
        },
        "required": [
          "file",
          "is_cloud",
          "discord_message_id"
        ],
        "title": "Body_upload_match_api_v2_matches_post",
        "type": "object"
      },
...
      "ContestBody": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ContestBody",
        "type": "object"
      },
      "ContestReport": {
        "properties": {
          "contestor_discord_id": {
            "title": "Contestor Discord Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "contestor_discord_id",
          "reason"
        ],
        "title": "ContestReport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeaderboardRankingResponse": {
        "properties": {
          "last_updated": {
            "title": "Last Updated",
            "type": "integer"
          },
          "rankings": {
            "items": {
              "$ref": "#/components/schemas/PlayerLeaderboard"
            },
            "title": "Rankings",
            "type": "array"
          }
        },
        "required": [
          "rankings",
          "last_updated"
        ],
        "title": "LeaderboardRankingResponse",
        "type": "object"
      },
...
      "MatchResponse": {
        "properties": {
          "affected_players": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AffectedPlayerRating"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Players"
          },
          "age": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approver_discord_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approver Discord Id"
          },
          "contest_report_list": {
            "items": {
              "$ref": "#/components/schemas/ContestReport"
            },
            "title": "Contest Report List",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discord_messages_id_list": {
            "items": {
              "type": "string"
            },
            "title": "Discord Messages Id List",
            "type": "array"
          },
          "flagged": {
            "title": "Flagged",
            "type": "boolean"
          },
          "flagged_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flagged By"
          },
          "game": {
            "title": "Game",
            "type": "string"
          },
          "game_mode": {
            "title": "Game Mode",
            "type": "string"
          },
          "is_cloud": {
            "title": "Is Cloud",
            "type": "boolean"
          },
          "map_type": {
            "title": "Map Type",
            "type": "string"
          },
          "match_id": {
            "title": "Match Id",
            "type": "string"
          },
          "parser_version": {
            "title": "Parser Version",
            "type": "string"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerSchema"
            },
            "title": "Players",
            "type": "array"
          },
          "reporter_discord_id": {
            "title": "Reporter Discord Id",
            "type": "string"
          },
          "save_file_hash": {
            "title": "Save File Hash",
            "type": "string"
          },
          "turn": {
            "title": "Turn",
            "type": "integer"
          }
        },
        "required": [
          "match_id",
          "game",
          "turn",
          "map_type",
          "game_mode",
          "is_cloud",
          "players",
          "parser_version",
          "discord_messages_id_list",
          "created_at",
          "flagged",
          "save_file_hash",
          "reporter_discord_id",
          "contest_report_list"
        ],
        "title": "MatchResponse",
        "type": "object"
      },
...
      "PlayerLeaderboard": {
        "properties": {
          "discord_id": {
            "title": "Discord Id",
            "type": "string"
          },
          "first": {
            "title": "First",
            "type": "integer"
          },
          "games_played": {
            "title": "Games Played",
            "type": "integer"
          },
          "rating": {
            "title": "Rating",
            "type": "integer"
          },
          "wins": {
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "discord_id",
          "rating",
          "games_played",
          "wins",
          "first"
        ],
        "title": "PlayerLeaderboard",
        "type": "object"
      },
      "PlayerSchema": {
        "properties": {
          "civ": {
            "title": "Civ",
            "type": "string"
          },
          "combined_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combined Delta"
          },
          "delta": {
            "default": 0.0,
            "title": "Delta",
            "type": "number"
          },
          "discord_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discord Id"
          },
          "is_sub": {
            "default": false,
            "title": "Is Sub",
            "type": "boolean"
          },
          "leader": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leader"
          },
          "placement": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement"
          },
          "player_alive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Alive"
          },
          "quit": {
            "title": "Quit",
            "type": "boolean"
          },
          "season_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Delta"
          },
          "steam_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steam Id"
          },
          "subbed_out": {
            "default": false,
            "title": "Subbed Out",
            "type": "boolean"
          },
          "team": {
            "title": "Team",
            "type": "integer"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          }
        },
        "required": [
          "civ",
          "team",
          "quit"
        ],
        "title": "PlayerSchema",
        "type": "object"
      },
      "PlayersPatch": {
        "properties": {
          "players": {
            "items": {
              "$ref": "#/components/schemas/SeatPatchIn"
            },
            "title": "Players",
            "type": "array"
          }
        },
        "title": "PlayersPatch",
        "type": "object"
      },
...
      "SeatPatchIn": {
        "description": "One seat's requested changes. Absent means unchanged.",
        "properties": {
          "discord_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discord Id"
          },
          "placement": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement"
          },
          "quit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quit"
          },
          "seat": {
            "title": "Seat",
            "type": "integer"
          },
          "sub_out": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Out"
          }
        },
        "required": [
          "seat"
        ],
        "title": "SeatPatchIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Civ Save Tool",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/matches": {
      "post": {
        "operationId": "upload_match_api_v2_matches_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-actor-discord-id",
            "required": true,
            "schema": {
              "title": "X-Actor-Discord-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_match_api_v2_matches_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Match",
        "tags": [
          "matches-v2"
        ]
      }
    },
    "/api/v2/matches/leaderboard": {
      "get": {
        "operationId": "get_leaderboard_api_v2_matches_leaderboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "game",
            "required": true,
            "schema": {
              "title": "Game",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "game_type",
            "required": true,
            "schema": {
              "title": "Game Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "game_mode",
            "required": true,
            "schema": {
              "title": "Game Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "is_seasonal",
            "required": false,
            "schema": {
              "default": false,
              "title": "Is Seasonal",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "is_combined",
            "required": false,
            "schema": {
              "default": false,
              "title": "Is Combined",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardRankingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Leaderboard",
        "tags": [
          "matches-v2"
        ]
      }
    },
    "/api/v2/matches/{match_id}": {
      "delete": {
        "operationId": "delete_match_api_v2_matches__match_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-actor-discord-id",
            "required": true,
            "schema": {
              "title": "X-Actor-Discord-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-is-staff",
            "required": false,
            "schema": {
              "default": false,
              "title": "X-Actor-Is-Staff",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Match",
        "tags": [
          "matches-v2"
        ]
      },
      "get": {
        "operationId": "get_match_api_v2_matches__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Match",
        "tags": [
          "matches-v2"
        ]
      }
    },
    "/api/v2/matches/{match_id}/approve": {
      "post": {
        "operationId": "approve_match_api_v2_matches__match_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-actor-discord-id",
            "required": true,
            "schema": {
              "title": "X-Actor-Discord-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Match",
        "tags": [
          "matches-v2"
        ]
      }
    },
    "/api/v2/matches/{match_id}/contest": {
      "post": {
        "operationId": "contest_match_api_v2_matches__match_id__contest_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-actor-discord-id",
            "required": true,
            "schema": {
              "title": "X-Actor-Discord-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-is-staff",
            "required": false,
            "schema": {
              "default": false,
              "title": "X-Actor-Is-Staff",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contest Match",
        "tags": [
          "matches-v2"
        ]
      }
    },
    "/api/v2/matches/{match_id}/players": {
      "patch": {
        "operationId": "patch_players_api_v2_matches__match_id__players_patch",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-actor-discord-id",
            "required": true,
            "schema": {
              "title": "X-Actor-Discord-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-is-staff",
            "required": false,
            "schema": {
              "default": false,
              "title": "X-Actor-Is-Staff",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayersPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Players",
        "tags": [
          "matches-v2"
        ]
      }
    },
    "/api/v2/matches/{match_id}/revert": {
      "post": {
        "operationId": "revert_match_api_v2_matches__match_id__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revert Match",
        "tags": [
          "matches-v2"
        ]
      }
    },
...
  }
}